+++ flowmachine/flowmachine/core/model_result.py
@@ -27,6 +27,7 @@ from flowmachine.core.errors.flowmachine_errors import (
 )
 from flowmachine.core.query_state import QueryStateMachine
 from flowmachine.core.query import Query
+from flowmachine.core.dependency_graph import store_all_unstored_dependencies
 
 import structlog
 
@@ -125,7 +126,12 @@ class ModelResult(Query):
         except AttributeError:
             return super().column_names
 
-    def to_sql(self, name: str, schema: Union[str, None] = None) -> Future:
+    def to_sql(
+        self,
+        name: str,
+        schema: Union[str, None] = None,
+        store_dependencies: bool = False,
+    ) -> Future:
         """
         Store the result of the calculation back into the database.
 
@@ -136,6 +142,8 @@ class ModelResult(Query):
         schema : str, default None
             Name of an existing schema. If none will use the postgres default,
             see postgres docs for more info.
+        store_dependencies : bool, default False
+            If True, store the dependencies of this query.
 
         Returns
         -------
@@ -155,6 +163,8 @@ class ModelResult(Query):
                 raise ValueError("Not computed yet.")
 
         def write_model_result(query_ddl_ops: List[str], connection: Engine) -> float:
+            if store_dependencies:
+                store_all_unstored_dependencies(self)
             self._df.to_sql(name, connection, schema=schema, index=False)
             QueryStateMachine(self.redis, self.query_id).finish()
             return self._runtime